Refactor code for main dropdown
